Modify boilerplate
The purpose of the typings declaration file is to tell TypeScript
that the Node module object has an id property. This is not related
to the development of the application, and should not be in the src
folder. Furthermore, this typing is only needed if Angular uses
module.id in a component decorator, and our use of Webpack makes it
unnecessary. Regardless, add a typing package for Node and remove
file.